+++ compiler/src/test/fixtures/operation_variables/output/java-cdi-client/com/example/operation_variables/EventService.java
@@ -1,0 +1,10 @@
+package com.example.operation_variables;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+
+@Generated("org.realityforge.giggle.Main")
+public interface EventService {
+  @Nonnull
+  EventQuery.Answer event(@Nonnull String id);
+}