+++ compiler/src/test/java/org/realityforge/giggle/MainTest.java
@@ -56,7 +56,13 @@ public class MainTest
                   "\n" +
                   "Supported Generators:\n" +
                   "  java-server\n" +
-                  "  java-client" );
+                  "  java-client\n" +
+                  "  java-cdi-client\n" +
+                  "   Supported Properties:\n" +
+                  "   - cdi.service.name (required): The name of the generated service class\n" +
+                  "   - cdi.base_url.jndi_name (required): The name of the JNDI resource that contains the base url for the endpoint\n" +
+                  "   - cdi.url.suffix: The path added to the setting retrieved from the base_url.config.key to construct the url. If unspecified then no suffix is added\n" +
+                  "   - cdi.keycloak.client.name: The name of the keycloak client used to authenticate the client. If unspecified then it is assumed no authentication step" );
   }
 
   @Test