+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.04](https://github.com/realityforge/giggle/tree/v0.04) (2019-06-17)
 [Full Changelog](https://github.com/realityforge/giggle/compare/v0.03...v0.04)