+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.guiceyloops` artifact to version `0.102`.
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.14`.
 
 ### [v0.05](https://github.com/realityforge/giggle/tree/v0.05) (2019-06-18)